Renamed org.cirdles.mars.cesar to org.cirdles.mars.sesar
Corrected the name of the package org.cirdles.mars.cesar to
org.cirdles.mars.sesar, as the package is meant to contain code relating
to the SESAR registry.
