--- sesar/src/main/java/org/cirdles/mars/cesar/jaxb/package-info.java
@@ -1,9 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2015.12.10 at 05:23:00 PM EST 
-//
-
-@javax.xml.bind.annotation.XmlSchema(namespace = "http://app.geosamples.org", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
-package org.cirdles.mars.cesar.jaxb;
