+++ sesar/src/main/java/org/cirdles/mars/sesar/SesarClient.java
@@ -1,8 +1,8 @@
-package org.cirdles.mars.cesar;
+package org.cirdles.mars.sesar;
 
 import org.cirdles.mars.api.SampleRegistryClient;
 import org.cirdles.mars.api.SampleSource;
-import org.cirdles.mars.cesar.jaxb.Samples;
+import org.cirdles.mars.sesar.jaxb.Samples;
 
 import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.client.Entity;
