+++ sesar/src/main/java/org/cirdles/mars/sesar/SesarException.java
@@ -1,4 +1,4 @@
-package org.cirdles.mars.cesar;
+package org.cirdles.mars.sesar;
 
 import org.cirdles.mars.api.MarsException;
