+++ sesar/src/main/java/org/cirdles/mars/sesar/jaxb/Classification.java
@@ -1,12 +1,12 @@
 //
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2015.12.10 at 05:23:00 PM EST 
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a>
+// Any modifications to this file will be lost upon recompilation of the source schema.
+// Generated on: 2015.12.10 at 05:23:00 PM EST
 //
 
 
-package org.cirdles.mars.cesar.jaxb;
+package org.cirdles.mars.sesar.jaxb;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -18,9 +18,9 @@ import javax.xml.bind.annotation.XmlType;
 
 /**
  * <p>Java class for anonymous complex type.
- * 
+ *
  * <p>The following schema fragment specifies the expected content contained within this class.
- * 
+ *
  * <pre>
  * &lt;complexType>
  *   &lt;complexContent>
@@ -131,8 +131,8 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;/complexContent>
  * &lt;/complexType>
  * </pre>
- * 
- * 
+ *
+ *
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
@@ -155,11 +155,11 @@ public class Classification {
 
     /**
      * Gets the value of the biology property.
-     * 
+     *
      * @return
      *     possible object is
      *     {@link Classification.Biology }
-     *     
+     *
      */
     public Classification.Biology getBiology() {
         return biology;
@@ -167,11 +167,11 @@ public class Classification {
 
     /**
      * Sets the value of the biology property.
-     * 
+     *
      * @param value
      *     allowed object is
      *     {@link Classification.Biology }
-     *     
+     *
      */
     public void setBiology(Classification.Biology value) {
         this.biology = value;
@@ -179,11 +179,11 @@ public class Classification {
 
     /**
      * Gets the value of the rock property.
-     * 
+     *
      * @return
      *     possible object is
      *     {@link Classification.Rock }
-     *     
+     *
      */
     public Classification.Rock getRock() {
         return rock;
@@ -191,11 +191,11 @@ public class Classification {
 
     /**
      * Sets the value of the rock property.
-     * 
+     *
      * @param value
      *     allowed object is
      *     {@link Classification.Rock }
-     *     
+     *
      */
     public void setRock(Classification.Rock value) {
         this.rock = value;
@@ -203,11 +203,11 @@ public class Classification {
 
     /**
      * Gets the value of the mineral property.
-     * 
+     *
      * @return
      *     possible object is
      *     {@link Classification.Mineral }
-     *     
+     *
      */
     public Classification.Mineral getMineral() {
         return mineral;
@@ -215,11 +215,11 @@ public class Classification {
 
     /**
      * Sets the value of the mineral property.
-     * 
+     *
      * @param value
      *     allowed object is
      *     {@link Classification.Mineral }
-     *     
+     *
      */
     public void setMineral(Classification.Mineral value) {
         this.mineral = value;
@@ -227,11 +227,11 @@ public class Classification {
 
     /**
      * Gets the value of the unknown property.
-     * 
+     *
      * @return
      *     possible object is
      *     {@link String }
-     *     
+     *
      */
     public String getUnknown() {
         return unknown;
@@ -239,11 +239,11 @@ public class Classification {
 
     /**
      * Sets the value of the unknown property.
-     * 
+     *
      * @param value
      *     allowed object is
      *     {@link String }
-     *     
+     *
      */
     public void setUnknown(String value) {
         this.unknown = value;
@@ -252,9 +252,9 @@ public class Classification {
 
     /**
      * <p>Java class for anonymous complex type.
-     * 
+     *
      * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
+     *
      * <pre>
      * &lt;complexType>
      *   &lt;complexContent>
@@ -277,8 +277,8 @@ public class Classification {
      *   &lt;/complexContent>
      * &lt;/complexType>
      * </pre>
-     * 
-     * 
+     *
+     *
      */
     @XmlAccessorType(XmlAccessType.FIELD)
     @XmlType(name = "", propOrder = {
@@ -294,11 +294,11 @@ public class Classification {
 
         /**
          * Gets the value of the macrobiology property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link Classification.Biology.Macrobiology }
-         *     
+         *
          */
         public Classification.Biology.Macrobiology getMacrobiology() {
             return macrobiology;
@@ -306,11 +306,11 @@ public class Classification {
 
         /**
          * Sets the value of the macrobiology property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link Classification.Biology.Macrobiology }
-         *     
+         *
          */
         public void setMacrobiology(Classification.Biology.Macrobiology value) {
             this.macrobiology = value;
@@ -318,11 +318,11 @@ public class Classification {
 
         /**
          * Gets the value of the microbiology property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link String }
-         *     
+         *
          */
         public String getMicrobiology() {
             return microbiology;
@@ -330,11 +330,11 @@ public class Classification {
 
         /**
          * Sets the value of the microbiology property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link String }
-         *     
+         *
          */
         public void setMicrobiology(String value) {
             this.microbiology = value;
@@ -343,9 +343,9 @@ public class Classification {
 
         /**
          * <p>Java class for anonymous complex type.
-         * 
+         *
          * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
+         *
          * <pre>
          * &lt;complexType>
          *   &lt;complexContent>
@@ -357,8 +357,8 @@ public class Classification {
          *   &lt;/complexContent>
          * &lt;/complexType>
          * </pre>
-         * 
-         * 
+         *
+         *
          */
         @XmlAccessorType(XmlAccessType.FIELD)
         @XmlType(name = "", propOrder = {
@@ -372,11 +372,11 @@ public class Classification {
 
             /**
              * Gets the value of the macrobiologyType property.
-             * 
+             *
              * @return
              *     possible object is
              *     {@link MacrobiologyDetails }
-             *     
+             *
              */
             public MacrobiologyDetails getMacrobiologyType() {
                 return macrobiologyType;
@@ -384,11 +384,11 @@ public class Classification {
 
             /**
              * Sets the value of the macrobiologyType property.
-             * 
+             *
              * @param value
              *     allowed object is
              *     {@link MacrobiologyDetails }
-             *     
+             *
              */
             public void setMacrobiologyType(MacrobiologyDetails value) {
                 this.macrobiologyType = value;
@@ -401,9 +401,9 @@ public class Classification {
 
     /**
      * <p>Java class for anonymous complex type.
-     * 
+     *
      * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
+     *
      * <pre>
      * &lt;complexType>
      *   &lt;complexContent>
@@ -415,8 +415,8 @@ public class Classification {
      *   &lt;/complexContent>
      * &lt;/complexType>
      * </pre>
-     * 
-     * 
+     *
+     *
      */
     @XmlAccessorType(XmlAccessType.FIELD)
     @XmlType(name = "", propOrder = {
@@ -429,11 +429,11 @@ public class Classification {
 
         /**
          * Gets the value of the mineralType property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link String }
-         *     
+         *
          */
         public String getMineralType() {
             return mineralType;
@@ -441,11 +441,11 @@ public class Classification {
 
         /**
          * Sets the value of the mineralType property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link String }
-         *     
+         *
          */
         public void setMineralType(String value) {
             this.mineralType = value;
@@ -456,9 +456,9 @@ public class Classification {
 
     /**
      * <p>Java class for anonymous complex type.
-     * 
+     *
      * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
+     *
      * <pre>
      * &lt;complexType>
      *   &lt;complexContent>
@@ -525,8 +525,8 @@ public class Classification {
      *   &lt;/complexContent>
      * &lt;/complexType>
      * </pre>
-     * 
-     * 
+     *
+     *
      */
     @XmlAccessorType(XmlAccessType.FIELD)
     @XmlType(name = "", propOrder = {
@@ -551,11 +551,11 @@ public class Classification {
 
         /**
          * Gets the value of the igneous property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link IgneousType }
-         *     
+         *
          */
         public IgneousType getIgneous() {
             return igneous;
@@ -563,11 +563,11 @@ public class Classification {
 
         /**
          * Sets the value of the igneous property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link IgneousType }
-         *     
+         *
          */
         public void setIgneous(IgneousType value) {
             this.igneous = value;
@@ -575,11 +575,11 @@ public class Classification {
 
         /**
          * Gets the value of the metamorphic property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link Classification.Rock.Metamorphic }
-         *     
+         *
          */
         public Classification.Rock.Metamorphic getMetamorphic() {
             return metamorphic;
@@ -587,11 +587,11 @@ public class Classification {
 
         /**
          * Sets the value of the metamorphic property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link Classification.Rock.Metamorphic }
-         *     
+         *
          */
         public void setMetamorphic(Classification.Rock.Metamorphic value) {
             this.metamorphic = value;
@@ -599,11 +599,11 @@ public class Classification {
 
         /**
          * Gets the value of the ore property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link Classification.Rock.Ore }
-         *     
+         *
          */
         public Classification.Rock.Ore getOre() {
             return ore;
@@ -611,11 +611,11 @@ public class Classification {
 
         /**
          * Sets the value of the ore property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link Classification.Rock.Ore }
-         *     
+         *
          */
         public void setOre(Classification.Rock.Ore value) {
             this.ore = value;
@@ -623,11 +623,11 @@ public class Classification {
 
         /**
          * Gets the value of the sedimentary property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link Classification.Rock.Sedimentary }
-         *     
+         *
          */
         public Classification.Rock.Sedimentary getSedimentary() {
             return sedimentary;
@@ -635,11 +635,11 @@ public class Classification {
 
         /**
          * Sets the value of the sedimentary property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link Classification.Rock.Sedimentary }
-         *     
+         *
          */
         public void setSedimentary(Classification.Rock.Sedimentary value) {
             this.sedimentary = value;
@@ -647,11 +647,11 @@ public class Classification {
 
         /**
          * Gets the value of the xenolithic property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link Classification.Rock.Xenolithic }
-         *     
+         *
          */
         public Classification.Rock.Xenolithic getXenolithic() {
             return xenolithic;
@@ -659,11 +659,11 @@ public class Classification {
 
         /**
          * Sets the value of the xenolithic property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link Classification.Rock.Xenolithic }
-         *     
+         *
          */
         public void setXenolithic(Classification.Rock.Xenolithic value) {
             this.xenolithic = value;
@@ -672,9 +672,9 @@ public class Classification {
 
         /**
          * <p>Java class for anonymous complex type.
-         * 
+         *
          * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
+         *
          * <pre>
          * &lt;complexType>
          *   &lt;complexContent>
@@ -686,8 +686,8 @@ public class Classification {
          *   &lt;/complexContent>
          * &lt;/complexType>
          * </pre>
-         * 
-         * 
+         *
+         *
          */
         @XmlAccessorType(XmlAccessType.FIELD)
         @XmlType(name = "", propOrder = {
@@ -701,11 +701,11 @@ public class Classification {
 
             /**
              * Gets the value of the metamorphicType property.
-             * 
+             *
              * @return
              *     possible object is
              *     {@link MetamorphicDetails }
-             *     
+             *
              */
             public MetamorphicDetails getMetamorphicType() {
                 return metamorphicType;
@@ -713,11 +713,11 @@ public class Classification {
 
             /**
              * Sets the value of the metamorphicType property.
-             * 
+             *
              * @param value
              *     allowed object is
              *     {@link MetamorphicDetails }
-             *     
+             *
              */
             public void setMetamorphicType(MetamorphicDetails value) {
                 this.metamorphicType = value;
@@ -728,9 +728,9 @@ public class Classification {
 
         /**
          * <p>Java class for anonymous complex type.
-         * 
+         *
          * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
+         *
          * <pre>
          * &lt;complexType>
          *   &lt;complexContent>
@@ -742,8 +742,8 @@ public class Classification {
          *   &lt;/complexContent>
          * &lt;/complexType>
          * </pre>
-         * 
-         * 
+         *
+         *
          */
         @XmlAccessorType(XmlAccessType.FIELD)
         @XmlType(name = "", propOrder = {
@@ -757,11 +757,11 @@ public class Classification {
 
             /**
              * Gets the value of the oreType property.
-             * 
+             *
              * @return
              *     possible object is
              *     {@link OreDetails }
-             *     
+             *
              */
             public OreDetails getOreType() {
                 return oreType;
@@ -769,11 +769,11 @@ public class Classification {
 
             /**
              * Sets the value of the oreType property.
-             * 
+             *
              * @param value
              *     allowed object is
              *     {@link OreDetails }
-             *     
+             *
              */
             public void setOreType(OreDetails value) {
                 this.oreType = value;
@@ -784,9 +784,9 @@ public class Classification {
 
         /**
          * <p>Java class for anonymous complex type.
-         * 
+         *
          * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
+         *
          * <pre>
          * &lt;complexType>
          *   &lt;complexContent>
@@ -798,8 +798,8 @@ public class Classification {
          *   &lt;/complexContent>
          * &lt;/complexType>
          * </pre>
-         * 
-         * 
+         *
+         *
          */
         @XmlAccessorType(XmlAccessType.FIELD)
         @XmlType(name = "", propOrder = {
@@ -813,11 +813,11 @@ public class Classification {
 
             /**
              * Gets the value of the sedimentaryType property.
-             * 
+             *
              * @return
              *     possible object is
              *     {@link SedimentaryDetails }
-             *     
+             *
              */
             public SedimentaryDetails getSedimentaryType() {
                 return sedimentaryType;
@@ -825,11 +825,11 @@ public class Classification {
 
             /**
              * Sets the value of the sedimentaryType property.
-             * 
+             *
              * @param value
              *     allowed object is
              *     {@link SedimentaryDetails }
-             *     
+             *
              */
             public void setSedimentaryType(SedimentaryDetails value) {
                 this.sedimentaryType = value;
@@ -840,9 +840,9 @@ public class Classification {
 
         /**
          * <p>Java class for anonymous complex type.
-         * 
+         *
          * <p>The following schema fragment specifies the expected content contained within this class.
-         * 
+         *
          * <pre>
          * &lt;complexType>
          *   &lt;complexContent>
@@ -865,8 +865,8 @@ public class Classification {
          *   &lt;/complexContent>
          * &lt;/complexType>
          * </pre>
-         * 
-         * 
+         *
+         *
          */
         @XmlAccessorType(XmlAccessType.FIELD)
         @XmlType(name = "", propOrder = {
@@ -882,11 +882,11 @@ public class Classification {
 
             /**
              * Gets the value of the igneous property.
-             * 
+             *
              * @return
              *     possible object is
              *     {@link IgneousType }
-             *     
+             *
              */
             public IgneousType getIgneous() {
                 return igneous;
@@ -894,11 +894,11 @@ public class Classification {
 
             /**
              * Sets the value of the igneous property.
-             * 
+             *
              * @param value
              *     allowed object is
              *     {@link IgneousType }
-             *     
+             *
              */
             public void setIgneous(IgneousType value) {
                 this.igneous = value;
@@ -906,11 +906,11 @@ public class Classification {
 
             /**
              * Gets the value of the metamorphic property.
-             * 
+             *
              * @return
              *     possible object is
              *     {@link Classification.Rock.Xenolithic.Metamorphic }
-             *     
+             *
              */
             public Classification.Rock.Xenolithic.Metamorphic getMetamorphic() {
                 return metamorphic;
@@ -918,11 +918,11 @@ public class Classification {
 
             /**
              * Sets the value of the metamorphic property.
-             * 
+             *
              * @param value
              *     allowed object is
              *     {@link Classification.Rock.Xenolithic.Metamorphic }
-             *     
+             *
              */
             public void setMetamorphic(Classification.Rock.Xenolithic.Metamorphic value) {
                 this.metamorphic = value;
@@ -931,9 +931,9 @@ public class Classification {
 
             /**
              * <p>Java class for anonymous complex type.
-             * 
+             *
              * <p>The following schema fragment specifies the expected content contained within this class.
-             * 
+             *
              * <pre>
              * &lt;complexType>
              *   &lt;complexContent>
@@ -945,8 +945,8 @@ public class Classification {
              *   &lt;/complexContent>
              * &lt;/complexType>
              * </pre>
-             * 
-             * 
+             *
+             *
              */
             @XmlAccessorType(XmlAccessType.FIELD)
             @XmlType(name = "", propOrder = {
@@ -960,11 +960,11 @@ public class Classification {
 
                 /**
                  * Gets the value of the metamorphicType property.
-                 * 
+                 *
                  * @return
                  *     possible object is
                  *     {@link MetamorphicDetails }
-                 *     
+                 *
                  */
                 public MetamorphicDetails getMetamorphicType() {
                     return metamorphicType;
@@ -972,11 +972,11 @@ public class Classification {
 
                 /**
                  * Sets the value of the metamorphicType property.
-                 * 
+                 *
                  * @param value
                  *     allowed object is
                  *     {@link MetamorphicDetails }
-                 *     
+                 *
                  */
                 public void setMetamorphicType(MetamorphicDetails value) {
                     this.metamorphicType = value;
