+++ sesar/src/main/java/org/cirdles/mars/sesar/jaxb/IgneousType.java
@@ -1,12 +1,12 @@
 //
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2015.12.10 at 05:23:00 PM EST 
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a>
+// Any modifications to this file will be lost upon recompilation of the source schema.
+// Generated on: 2015.12.10 at 05:23:00 PM EST
 //
 
 
-package org.cirdles.mars.cesar.jaxb;
+package org.cirdles.mars.sesar.jaxb;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -17,9 +17,9 @@ import javax.xml.bind.annotation.XmlType;
 
 /**
  * <p>Java class for IgneousType complex type.
- * 
+ *
  * <p>The following schema fragment specifies the expected content contained within this class.
- * 
+ *
  * <pre>
  * &lt;complexType name="IgneousType">
  *   &lt;complexContent>
@@ -52,8 +52,8 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;/complexContent>
  * &lt;/complexType>
  * </pre>
- * 
- * 
+ *
+ *
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "IgneousType", propOrder = {
@@ -69,11 +69,11 @@ public class IgneousType {
 
     /**
      * Gets the value of the plutonic property.
-     * 
+     *
      * @return
      *     possible object is
      *     {@link IgneousType.Plutonic }
-     *     
+     *
      */
     public IgneousType.Plutonic getPlutonic() {
         return plutonic;
@@ -81,11 +81,11 @@ public class IgneousType {
 
     /**
      * Sets the value of the plutonic property.
-     * 
+     *
      * @param value
      *     allowed object is
      *     {@link IgneousType.Plutonic }
-     *     
+     *
      */
     public void setPlutonic(IgneousType.Plutonic value) {
         this.plutonic = value;
@@ -93,11 +93,11 @@ public class IgneousType {
 
     /**
      * Gets the value of the volcanic property.
-     * 
+     *
      * @return
      *     possible object is
      *     {@link IgneousType.Volcanic }
-     *     
+     *
      */
     public IgneousType.Volcanic getVolcanic() {
         return volcanic;
@@ -105,11 +105,11 @@ public class IgneousType {
 
     /**
      * Sets the value of the volcanic property.
-     * 
+     *
      * @param value
      *     allowed object is
      *     {@link IgneousType.Volcanic }
-     *     
+     *
      */
     public void setVolcanic(IgneousType.Volcanic value) {
         this.volcanic = value;
@@ -118,9 +118,9 @@ public class IgneousType {
 
     /**
      * <p>Java class for anonymous complex type.
-     * 
+     *
      * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
+     *
      * <pre>
      * &lt;complexType>
      *   &lt;complexContent>
@@ -132,8 +132,8 @@ public class IgneousType {
      *   &lt;/complexContent>
      * &lt;/complexType>
      * </pre>
-     * 
-     * 
+     *
+     *
      */
     @XmlAccessorType(XmlAccessType.FIELD)
     @XmlType(name = "", propOrder = {
@@ -147,11 +147,11 @@ public class IgneousType {
 
         /**
          * Gets the value of the plutonicType property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link IgneousDetails }
-         *     
+         *
          */
         public IgneousDetails getPlutonicType() {
             return plutonicType;
@@ -159,11 +159,11 @@ public class IgneousType {
 
         /**
          * Sets the value of the plutonicType property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link IgneousDetails }
-         *     
+         *
          */
         public void setPlutonicType(IgneousDetails value) {
             this.plutonicType = value;
@@ -174,9 +174,9 @@ public class IgneousType {
 
     /**
      * <p>Java class for anonymous complex type.
-     * 
+     *
      * <p>The following schema fragment specifies the expected content contained within this class.
-     * 
+     *
      * <pre>
      * &lt;complexType>
      *   &lt;complexContent>
@@ -188,8 +188,8 @@ public class IgneousType {
      *   &lt;/complexContent>
      * &lt;/complexType>
      * </pre>
-     * 
-     * 
+     *
+     *
      */
     @XmlAccessorType(XmlAccessType.FIELD)
     @XmlType(name = "", propOrder = {
@@ -203,11 +203,11 @@ public class IgneousType {
 
         /**
          * Gets the value of the volcanicType property.
-         * 
+         *
          * @return
          *     possible object is
          *     {@link IgneousDetails }
-         *     
+         *
          */
         public IgneousDetails getVolcanicType() {
             return volcanicType;
@@ -215,11 +215,11 @@ public class IgneousType {
 
         /**
          * Sets the value of the volcanicType property.
-         * 
+         *
          * @param value
          *     allowed object is
          *     {@link IgneousDetails }
-         *     
+         *
          */
         public void setVolcanicType(IgneousDetails value) {
             this.volcanicType = value;
