+++ sesar/src/main/java/org/cirdles/mars/sesar/jaxb/ObjectFactory.java
@@ -6,7 +6,7 @@
 //
 
 
-package org.cirdles.mars.cesar.jaxb;
+package org.cirdles.mars.sesar.jaxb;
 
 import javax.xml.bind.JAXBElement;
 import javax.xml.bind.annotation.XmlElementDecl;
@@ -22,7 +22,7 @@ import java.math.BigInteger;
 /**
  * This object contains factory methods for each
  * Java content interface and Java element interface
- * generated in the org.cirdles.mars.cesar.jaxb package.
+ * generated in the org.cirdles.mars.sesar.jaxb package.
  * <p>An ObjectFactory allows you to programatically
  * construct new instances of the Java representation
  * for XML content. The Java representation of XML
@@ -103,7 +103,7 @@ public class ObjectFactory {
     private final static QName _Locality_QNAME = new QName("http://app.geosamples.org", "locality");
 
     /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: org.cirdles.mars.cesar.jaxb
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: org.cirdles.mars.sesar.jaxb
      *
      */
     public ObjectFactory() {
