+++ sesar/src/test/java/org/cirdles/mars/sesar/SesarTranslatorTest.java
@@ -1,7 +1,7 @@
-package org.cirdles.mars.cesar;
+package org.cirdles.mars.sesar;
 
 import org.cirdles.mars.api.Sample;
-import org.cirdles.mars.cesar.jaxb.Samples;
+import org.cirdles.mars.sesar.jaxb.Samples;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
